+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - New `formatDate` extension added over DateTime class
 - Added some extensions for String class
 - Added extension methods for ConnectionState in Flutter widgets.
+- Refactor dialog box and bottom sheet functions to accept a function that
+  returns a widget.
 - [Deprecated]
   - `pr` getter extension has been deprecated.
 - Improved pub points